move internal types to the separate files
