+++ plugins/org.eclipse.n4js.generator.headless/src/org/eclipse/n4js/generator/headless/N4HeadlessCompiler.java
@@ -17,10 +17,8 @@ import java.text.DecimalFormat;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.LinkedHashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.ListIterator;
@@ -111,14 +109,6 @@ public class N4HeadlessCompiler {
 
 	/** Abstraction to the file system, used by the generators */
 	private final JavaIoFileSystemAccess fsa;
-
-	/** Compares two N4JS projects. Used for sorting and comparing project dependencies. */
-	private final static Comparator<IN4JSProject> n4JSProjComparator = new Comparator<IN4JSProject>() {
-		@Override
-		public int compare(IN4JSProject o1, IN4JSProject o2) {
-			return o1.getProjectId().compareTo(o2.getProjectId());
-		}
-	};
 
 	/** N4JS-Implementation of a workspace without OSGI */
 	@Inject
@@ -383,41 +373,6 @@ public class N4HeadlessCompiler {
 	}
 
 	/**
-	 * Encapsulates the result of {@link N4HeadlessCompiler#collectAndRegisterProjects(List, List, List)}.
-	 */
-	private static class BuildSet {
-
-		/**
-		 * The projects which the user explicitly requested to be compiled. If the user requested compilation of
-		 * specific single files, then this list contains the projects containing the files.
-		 */
-		public final List<N4JSProject> requestedProjects;
-
-		/**
-		 * The projects which were discovered as dependencies of the above projects, without having been requested to be
-		 * compiled by the user. In other words, these projects are only being compiled because a requested project
-		 * depends on them.
-		 */
-		public final List<N4JSProject> discoveredProjects;
-
-		/**
-		 * A predicate that indicates whether or not a given resource, identified by its URI, should be processed. If
-		 * the user requested compilation of specific single files, then this predicate applies only to those files, and
-		 * no others. In all other cases, the predicate applies to every file, i.e., it always returns
-		 * <code>true</code>.
-		 */
-		public final Predicate<URI> resourceFilter;
-
-		public BuildSet(List<N4JSProject> requestedProjects, List<N4JSProject> discoveredProjects,
-				Predicate<URI> projectFilter) {
-			this.requestedProjects = requestedProjects;
-			this.discoveredProjects = discoveredProjects;
-			this.resourceFilter = projectFilter;
-		}
-
-	}
-
-	/**
 	 * Clean the output folders of all <b>direct</b> projects found in search paths.
 	 *
 	 * @param searchPaths
@@ -676,13 +631,15 @@ public class N4HeadlessCompiler {
 		allProjectsToCompile.stream().forEach(project -> markedProjects.put(project, new MarkedProject(project)));
 
 		// Maps a project to the projects that depend on it.
-		Multimap<IN4JSProject, IN4JSProject> pendencies = TreeMultimap.create(n4JSProjComparator, n4JSProjComparator);
+		Multimap<IN4JSProject, IN4JSProject> pendencies = TreeMultimap.create(N4JSProjectComparator.INSTANCE,
+				N4JSProjectComparator.INSTANCE);
 
 		// Maps a project to the projects it depends on.
-		Multimap<IN4JSProject, IN4JSProject> dependencies = TreeMultimap.create(n4JSProjComparator, n4JSProjComparator);
+		Multimap<IN4JSProject, IN4JSProject> dependencies = TreeMultimap.create(N4JSProjectComparator.INSTANCE,
+				N4JSProjectComparator.INSTANCE);
 
 		// Sorted set of projects without dependencies (starting points) to determine their order.
-		SortedSet<IN4JSProject> independentProjects = new TreeSet<>(n4JSProjComparator);
+		SortedSet<IN4JSProject> independentProjects = new TreeSet<>(N4JSProjectComparator.INSTANCE);
 
 		// Initialize preconditions, dependencies, and independent projects.
 		computeDependencyGraph(markedProjects.keySet(), pendencies, dependencies, independentProjects);
@@ -1649,193 +1606,5 @@ public class N4HeadlessCompiler {
 	 */
 	public void setLogFile(String logFile) {
 		this.logFile = logFile;
-	}
-
-	/*
-	 * ===============================================================================================================
-	 *
-	 * MARKED PROJECT UTILITY CLASS
-	 *
-	 * ===============================================================================================================
-	 */
-	/**
-	 * A wrapper around N4JS projects that has the ability to track dependent projects in the form of markers. A project
-	 * is added as a marker of this project if it has an active dependency on it. A dependency is active as long as the
-	 * dependent project is not yet built.
-	 * <p>
-	 * Additionally, projects that have been explicitly requested to be compiled by the user are added as markers to
-	 * themselves, as opposed to projects that have been discovered solely due to dependencies of explicitly requested
-	 * projects.
-	 * </p>
-	 * <p>
-	 * Furthermore, this class tracks the loaded resources of the wrapped projects in order to be able to unload them as
-	 * soon as possible.
-	 * </p>
-	 */
-	static class MarkedProject {
-		/**
-		 * The wrapped project.
-		 */
-		final IN4JSProject project;
-
-		/**
-		 * Contains the active markers, i.e., dependent projects that have not yet been built themselves.
-		 */
-		final Set<IN4JSProject> markers = new TreeSet<>(n4JSProjComparator);
-
-		/**
-		 * All loaded resources of this project.
-		 */
-		final Set<Resource> resources = new LinkedHashSet<>();
-
-		/**
-		 * All loaded external resources of this project. This is a subset of {@link #resources}.
-		 */
-		final Set<Resource> externalResources = new HashSet<>();
-
-		/**
-		 * All loaded test resources of this project. This is a subset of {@link #resources}.
-		 */
-		final Set<Resource> testResources = new HashSet<>();
-
-		/**
-		 * Create a wrapper around a project;
-		 */
-		public MarkedProject(IN4JSProject project) {
-			this.project = project;
-		}
-
-		/**
-		 * Indicates whether the given resource is external in the context of this project.
-		 *
-		 * @param resource
-		 *            the resource to check
-		 * @return <code>true</code> if the given resource is external and <code>false</code> otherwise
-		 */
-		public boolean isExternal(Resource resource) {
-			return externalResources.contains(resource);
-		}
-
-		/**
-		 * Indicates whether the given resource is a test in the context of this project.
-		 *
-		 * @param resource
-		 *            the resource to check
-		 * @return <code>true</code> if the given resource is a test and <code>false</code> otherwise
-		 */
-		public boolean isTest(Resource resource) {
-			return testResources.contains(resource);
-		}
-
-		/**
-		 * Adds the given project as a marker, indicating that it depends on this project.
-		 *
-		 * @param marker
-		 *            the project to mark this project with
-		 */
-		public void markWith(IN4JSProject marker) {
-			markers.add(marker);
-		}
-
-		/**
-		 * Indicates whether or not the given project is a marker of this project.
-		 *
-		 * @param marker
-		 *            the project to check
-		 * @return <code>true</code> if the given project is a marker of this project and <code>false</code> otherwise
-		 */
-		public boolean hasMarker(IN4JSProject marker) {
-			return markers.contains(marker);
-		}
-
-		/**
-		 * Indicates whether or not this project still has markers.
-		 *
-		 * @return <code>true</code> if this project still has markers and <code>false</code> otherwise
-		 */
-		public boolean hasMarkers() {
-			return !markers.isEmpty();
-		}
-
-		/**
-		 * Remove the given project as a marker of this project, indicating that it is no longer has an active
-		 * dependency on this project.
-		 *
-		 * @param marker
-		 *            dependent project to be removed
-		 * @return <code>true</code> if the given project was a marker of this project and <code>false</code> otherwise
-		 */
-		public boolean remove(IN4JSProject marker) {
-			return markers.remove(marker);
-		}
-
-		/**
-		 * Unload all resources associated with this marked project and remove them from the given resource set.
-		 *
-		 * @param resourceSet
-		 *            the resource set containing the resources of this project
-		 * @param recorder
-		 *            the progress state recorder
-		 */
-		public void unload(ResourceSet resourceSet, N4ProgressStateRecorder recorder) {
-			recorder.markStartUnloading(this);
-
-			ResourceDescriptionsData index = ResourceDescriptionsData.ResourceSetAdapter
-					.findResourceDescriptionsData(resourceSet);
-
-			unloadResources(resourceSet, index, recorder);
-			unloadManifestResource(resourceSet, index, recorder);
-			clearResources();
-
-			recorder.markFinishedUnloading(this);
-		}
-
-		private void unloadResources(ResourceSet resourceSet, ResourceDescriptionsData index,
-				N4ProgressStateRecorder recorder) {
-			for (Resource res : resources)
-				unloadResource(res, resourceSet, index, recorder);
-		}
-
-		private void unloadManifestResource(ResourceSet resourceSet, ResourceDescriptionsData index,
-				N4ProgressStateRecorder recorder) {
-			Optional<URI> manifestLocation = project.getManifestLocation();
-			if (manifestLocation.isPresent()) {
-				Resource resource = resourceSet.getResource(manifestLocation.get(), false);
-				if (resource != null)
-					unloadResource(resource, resourceSet, index, recorder);
-			}
-		}
-
-		private void unloadResource(Resource resource, ResourceSet resourceSet, ResourceDescriptionsData index,
-				N4ProgressStateRecorder recorder) {
-			recorder.markUnloadingOf(resource);
-			if (index != null)
-				index.removeDescription(resource.getURI());
-			resource.unload();
-			resourceSet.getResources().remove(resource);
-		}
-
-		private void clearResources() {
-			resources.clear();
-			externalResources.clear();
-			testResources.clear();
-		}
-
-		/**
-		 * Unload the ASTs and clear the resource scope caches of all resources that belong to this marked project.
-		 */
-		public void unloadASTAndClearCaches() {
-			Iterables.filter(resources, resource -> resource.isLoaded()).forEach(resource -> {
-				if (resource instanceof N4JSResource) {
-					N4JSResource n4jsResource = (N4JSResource) resource;
-					n4jsResource.unloadAST();
-				}
-			});
-		}
-
-		@Override
-		public String toString() {
-			return project.toString();
-		}
 	}
 }
